Small cleanup in Bolt module
1) moved classes used only in tests to the `test` folder
2) extracted couple nested classes to the top level
3) replaced direct buffer with heap buffer in `ProtocolChooser` to make it's
 garbage collection more predictable
